Coursework 1, Exercise 4
The Task:

Write a function that implements the bubble sort method for an array of integers. 
The declaration of this function must be: int* Bubble(int, int *);
Here the first parameter is the length of array and the second is the pointer to the array. 
The function returns pointer to the sorted array (must be a new array, the input array must stay as it is).
To test the function write main that reads the array length, fills the array with pseudo-random integers, calls Bubble and prints the sorted array.